hash and store

+++ download_models.sh
@@ -1,4 +1,8 @@
 #!/usr/bin/env bash
-echo "Downloading yolov3 weights"
-mkdir weights
-wget -O yolo/yolov3.weights https://pjreddie.com/media/files/yolov3.weights
+if [ ! -f ./yolo/yolov3.weights ]; then
+    echo "YOLO Weights not found! Downloading ..."
+	wget -O yolo/yolov3.weights https://pjreddie.com/media/files/yolov3.weights
+fi
+
+docker-compose build
+docker-compose up -d
